requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
aiogram==3.13.1
APScheduler==3.10.4
Jinja2==3.1.4
python-multipart==0.0.17
